DTO 수정

+++ Angbit/src/main/java/com/oracle/Angbit/model/common/Coin.java
@@ -9,7 +9,7 @@ public class Coin {
 	
 	private String coincode;
 	private String id;
-	private int coin_amt;
+	private Float coin_amt;
 	private int avg_price;
 	private int tot_price;
